move LineStyleLayer tests to their own file
